Move PoolType type to pool types file
