+++ src/composables/pools/usePoolCreation.ts
@@ -11,8 +11,8 @@ import useTransactions from '@/composables/useTransactions';
 import { POOLS } from '@/constants/pools';
 import { bnum, lsRemove, lsSet, scale } from '@/lib/utils';
 import { balancerService } from '@/services/balancer/balancer.service';
-import { PoolType } from '@/services/balancer/subgraph/types';
 import { configService } from '@/services/config/config.service';
+import { PoolType } from '@/services/pool/types';
 import useWeb3 from '@/services/web3/useWeb3';
 
 import useTokens from '../useTokens';
